+++ src/main/kotlin/veridius/discover/services/connection/internal/MySQLConnection.kt
@@ -1,7 +1,5 @@
-package veridius.discover.services.connection
+package veridius.discover.services.connection.internal
 
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.withContext
 import mu.KotlinLogging
 import org.springframework.boot.jdbc.DataSourceBuilder
 import veridius.discover.configuration.properties.DatabaseConfigurationProperties.DatabaseConnectionConfiguration
