+++ src/main/kotlin/veridius/discover/services/connection/internal/PostgresConnection.kt
@@ -1,11 +1,8 @@
-package veridius.discover.services.connection
+package veridius.discover.services.connection.internal
 
 import com.zaxxer.hikari.HikariConfig
 import com.zaxxer.hikari.HikariDataSource
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.withContext
 import mu.KotlinLogging
-import org.springframework.boot.jdbc.DataSourceBuilder
 import veridius.discover.configuration.properties.DatabaseConfigurationProperties.DatabaseConnectionConfiguration
 import javax.sql.DataSource
