--- src/main/kotlin/veridius/discover/services/health/DatabaseHealthChecker.kt
@@ -1,4 +1,0 @@
-package veridius.discover.services.health
-
-class DatabaseHealthChecker {
-}
