+++ src/main/kotlin/veridius/discover/services/health/HealthService.kt
@@ -1,0 +1,7 @@
+package veridius.discover.services.health
+
+import org.springframework.stereotype.Service
+
+@Service
+class HealthService {
+}
