+++ src/main/kotlin/veridius/discover/services/metrics/MetricsService.kt
@@ -1,0 +1,7 @@
+package veridius.discover.services.metrics
+
+import org.springframework.stereotype.Service
+
+@Service
+class MetricsService {
+}
