+++ src/main/kotlin/veridius/discover/services/monitoring/MonitoringService.kt
@@ -1,7 +1,8 @@
 package veridius.discover.services.monitoring
 
 import org.springframework.stereotype.Component
+import org.springframework.stereotype.Service
 
-@Component
-class DatabaseMonitoringManager {
+@Service
+class MonitoringService {
 }
