+++ src/main/resources/datasource.yml
@@ -1,0 +1,5 @@
+spring:
+  datasource:
+    url: ${POSTGRES_JDBC}
+    username: ${POSTGRES_USER}
+    password: ${POSTGRES_PASSWORD}
