+++ src/main/resources/kafka.yml
@@ -1,0 +1,6 @@
+spring:
+  kafka:
+    bootstrap-servers: localhost:9092
+    consumer:
+      group-id: event-discovery
+      auto-offset-reset: earliest
